Time: 1011 ms (70.80%), Space: 305.4 MB (8.56%) - LeetHub

+++ 3164-find-the-number-of-good-pairs-ii/3164-find-the-number-of-good-pairs-ii.cpp
@@ -1,5 +1,5 @@
 using ll = long long;
-const ll sz = 1e6;
+const int sz = 1e6;
 class Solution {
 public:
     long long numberOfPairs(vector<int>& a, vector<int>& b, int k) {
